Separates form into its own component
Also, moves addItem into the agent.js
